requirements.txt
# Core ML
torch==2.6.0
torchvision==0.21.0
numpy==2.2.6

# Image processing
Pillow==12.0.0
opencv-python==4.8.1.78

# Web framework
fastapi==0.104.1
uvicorn==0.24.0

# COCO dataset tools
pycocotools==2.0.7

# Data handling
scikit-learn==1.3.2
scipy==1.11.4

# Visualization
matplotlib==3.8.2

# Utilities
python-multipart==0.0.6
tqdm==4.66.1
